image-classifier-python/requirements.txt
flask==1.1.2
gevent==20.5.0
requests==2.23.0
tensorflow==2.2.0
pillow==7.1.2